remove repo filtering on the github credentials
The github-org processor does not use the concept of repositories.
As such filtering based on selected repositories does not make sense
for that processer. The GithubCredentials was imposing what appears
like an unnessesary restriction on the tokens using the repo name.

By removing this restriction, it enalbes the github-org processor to
work when a github app installation has a repository restriction.

Signed-off-by: Brian Fletcher <brian@roadie.io>